+++ packages/core-js/actual/iterator/concat.js
@@ -2,6 +2,7 @@
 var parent = require('../../stable/iterator/concat');
 require('../../modules/esnext.iterator.chunks');
 require('../../modules/esnext.iterator.concat');
+require('../../modules/esnext.iterator.includes');
 require('../../modules/esnext.iterator.windows');
 
 module.exports = parent;