+++ packages/core-js/actual/iterator/zip.js
@@ -1,6 +1,7 @@
 'use strict';
 var parent = require('../../stable/iterator/zip');
 require('../../modules/esnext.iterator.chunks');
+require('../../modules/esnext.iterator.includes');
 require('../../modules/esnext.iterator.windows');
 require('../../modules/esnext.iterator.zip');