+++ packages/core-js/proposals/iterator-includes.js
@@ -1,0 +1,3 @@
+'use strict';
+// https://github.com/tc39/proposal-iterator-includes
+require('../modules/esnext.iterator.includes');