+++ tests/eslint/eslint.config.js
@@ -1300,6 +1300,7 @@ const forbidCompletelyNonExistentBuiltIns = {
   ] }],
   'es/no-nonstandard-iterator-prototype-properties': [ERROR, { allow: [
     'chunks',
+    'includes',
     'sliding',
     'toAsync',
     'windows',